+++ nnunet/training/loss_functions/TopK_loss.py
@@ -28,6 +28,6 @@ class TopKLoss(CrossentropyND):
     def forward(self, inp, target):
         target = target[:, 0].long()
         res = super(TopKLoss, self).forward(inp, target)
-        num_voxels = np.prod(res.shape)
+        num_voxels = np.prod(res.shape, dtype=np.int64)
         res, _ = torch.topk(res.view((-1, )), int(num_voxels * self.k / 100), sorted=False)
         return res.mean()
